+++ src/Server.h
@@ -216,28 +216,44 @@ public:
 
   /**
    * @brief Get a list of all Stations
-   * @return vector with object stationer
+   * @return vector with station objects
    */
   Object::StationVector getStations() const;
 
   /**
-   * @brief Get a Station that exists at this NetworkStation and is identified
-   * via information object address
-   * @return Stationer to Station or nullptr
+   * @brief Retrieves a Station that exists on this Server, identified by the
+   * given information object address.
+   * @param ca The common address (CA) that uniquely identifies the Station.
+   * @return A shared pointer to the Station if it exists, or nullptr if no
+   * matching Station is found.
    */
   std::shared_ptr<Object::Station>
   getStation(std::uint_fast16_t commonAddress) const;
 
   /**
-   * @brief Test if Stations exists at this NetworkStation
-   * @return information on availability of child Station objects
+   * @brief Checks whether a Station with the given common address exists on
+   * this Server.
+   * @param commonAddress The common address (CA) used to identify the Station.
+   * @return True if a Station with the specified common address exists,
+   * otherwise false.
    */
   bool hasStation(std::uint_fast16_t commonAddress) const;
 
   /**
-   * @brief Add a Station to this Station
+   * @brief Adds a new Station to this Server.
+   * @param commonAddress The common address (CA) that uniquely identifies the
+   * new Station.
+   * @return A shared pointer to the newly added Station.
    */
   std::shared_ptr<Object::Station> addStation(std::uint_fast16_t commonAddress);
+
+  /**
+   * @brief Removes an existing Station from this Server.
+   *
+   * @param commonAddress The common address (CA) of the Station to be removed.
+   * @return True if the Station was successfully removed, otherwise false.
+   */
+  bool removeStation(std::uint_fast16_t commonAddress);
 
   /**
    * @brief Get a reference to the protocol parameters to be able to read and